+++ ViewModels/StudentViewModel.cs
@@ -14,6 +14,6 @@ namespace ViewModels
         [Display(Name="Full Name")]
         public string Name { get; set; }
         public int? Age { get; set; }
-        public int? Gender { get; set; }
+        public byte? Gender { get; set; }
     }
 }
